Uncouple treebuilding from AdvancedTreeSearch
