+++ apptainer/2022-10-21_tensorflow-1.15_arm_v1/makefiles/Modules.make
@@ -145,6 +145,7 @@ endif
 
 # ****** Libraries ******
 LIBS_SEARCH = src/Search/libSprintSearch.$(a)
+LIBS_SEARCH += src/Search/TreeBuilder/libSprintTreeBuilder.$(a)
 ifdef MODULE_SEARCH_WFST
 LIBS_SEARCH += src/Search/Wfst/libSprintSearchWfst.$(a)
 LIBS_SEARCH += src/OpenFst/libSprintOpenFst.$(a)
