+++ src/Search/AdvancedTreeSearch/SearchSpace.cc
@@ -23,13 +23,14 @@
 #include <Lm/Module.hh>
 #include <Mm/GaussDiagonalMaximumFeatureScorer.hh>
 
+#include <Search/TreeBuilder/PersistentStateTree.hh>
+#include <Search/TreeBuilder/TreeBuilder.hh>
+
 #include "AcousticLookAhead.hh"
-#include "PersistentStateTree.hh"
 #include "PrefixFilter.hh"
 #include "Pruning.hh"
 #include "SearchNetworkTransformation.hh"
 #include "SearchSpaceStatistics.hh"
-#include "TreeBuilder.hh"
 
 using namespace AdvancedTreeSearch;
 
@@ -132,16 +133,16 @@ const Core::ParameterBool paramBuildMinimizedTreeFromScratch(
         true);
 
 const Core::Choice choiceTreeBuilderType(
-        "classic-hmm", static_cast<int>(StaticSearchAutomaton::TreeBuilderType::classicHmm),
-        "minimized-hmm", static_cast<int>(StaticSearchAutomaton::TreeBuilderType::minimizedHmm),
-        "ctc", static_cast<int>(StaticSearchAutomaton::TreeBuilderType::ctc),
+        "classic-hmm", static_cast<int>(TreeBuilderType::classicHmm),
+        "minimized-hmm", static_cast<int>(TreeBuilderType::minimizedHmm),
+        "ctc", static_cast<int>(TreeBuilderType::ctc),
         Core::Choice::endMark());
 
 const Core::ParameterChoice paramTreeBuilderType(
         "tree-builder-type",
         &choiceTreeBuilderType,
         "which tree builder to use",
-        static_cast<int>(StaticSearchAutomaton::TreeBuilderType::previousBehavior));
+        static_cast<int>(TreeBuilderType::previousBehavior));
 
 const Core::ParameterBool paramConditionPredecessorWord(
         "condition-on-predecessor-word",
@@ -386,7 +387,7 @@ StaticSearchAutomaton::StaticSearchAutomaton(Core::Configuration config, Core::R
         : Precursor(config),
           hmmLength(acousticModel->hmmTopologySet()->getDefault().nPhoneStates() * acousticModel->hmmTopologySet()->getDefault().nSubStates()),
           minimized(paramBuildMinimizedTreeFromScratch(config)),
-          network(config, acousticModel, lexicon, std::bind(&StaticSearchAutomaton::createTreeBuilder, this, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3, std::placeholders::_4, std::placeholders::_5)),
+          network(config, acousticModel, lexicon, std::bind(&createTreeBuilder, std::placeholders::_1, std::placeholders::_2, std::placeholders::_3, std::placeholders::_4, std::placeholders::_5, std::placeholders::_6)),
           prefixFilter(nullptr),
           treeBuilderType_(static_cast<TreeBuilderType>(paramTreeBuilderType(config))),
           acousticModel_(acousticModel),
@@ -408,7 +409,7 @@ void StaticSearchAutomaton::buildNetwork() {
     if (!network.read(transformation)) {
         log() << "persistent network image could not be loaded, building it";
 
-        std::unique_ptr<AbstractTreeBuilder> builder = createTreeBuilder(config, *lexicon_, *acousticModel_, network);
+        std::unique_ptr<AbstractTreeBuilder> builder = createTreeBuilder(treeBuilderType_, config, *lexicon_, *acousticModel_, network);
         if (not builder) {
             network.build();
             network.cleanup();
@@ -766,21 +767,6 @@ void StaticSearchAutomaton::buildBatches() {
     log() << "number of pushed labels: " << pushedLabels << " unpushed: " << unpushedLabels;
 
     network.removeOutputs();
-}
-
-std::unique_ptr<AbstractTreeBuilder> StaticSearchAutomaton::createTreeBuilder(Core::Configuration config, const Bliss::Lexicon& lexicon, const Am::AcousticModel& acousticModel, Search::PersistentStateTree& network, bool initialize) {
-    switch (treeBuilderType_) {
-        case TreeBuilderType::classicHmm: {  // Use StateTree.hh
-            return std::unique_ptr<AbstractTreeBuilder>(nullptr);
-        } break;
-        case TreeBuilderType::minimizedHmm: {  // Use TreeStructure.hh
-            return std::unique_ptr<AbstractTreeBuilder>(new MinimizedTreeBuilder(config, *lexicon_, *acousticModel_, network, initialize));
-        } break;
-        case TreeBuilderType::ctc: {
-            return std::unique_ptr<AbstractTreeBuilder>(new CtcTreeBuilder(config, *lexicon_, *acousticModel_, network, initialize));
-        } break;
-        default: defect();
-    }
 }
 
 // ------------------------------- Search Space --------------------------------
