+++ src/Search/Makefile
@@ -15,8 +15,6 @@ LIBSPRINTSEARCH_O	= \
 		$(OBJDIR)/LanguageModelLookahead.o \
 		$(OBJDIR)/Module.o \
 		$(OBJDIR)/Search.o \
-		$(OBJDIR)/StateTree.o \
-		$(OBJDIR)/StateTreeIo.o \
 		$(OBJDIR)/WordConditionedTreeSearch.o
 
 CHECK_O		= $(OBJDIR)/check.o \
@@ -24,7 +22,6 @@ CHECK_O		= $(OBJDIR)/check.o \
 			  ../Bliss/libSprintBliss.$(a) \
 			  ../Fsa/libSprintFsa.$(a) \
 			  ../Core/libSprintCore.$(a)
-			  
 
 ifdef MODULE_SEARCH_MBR
 LIBSPRINTSEARCH_O += $(OBJDIR)/MinimumBayesRiskSearch.o
@@ -32,6 +29,7 @@ LIBSPRINTSEARCH_O += $(OBJDIR)/MinimumBayesRiskAStarSearch.o
 LIBSPRINTSEARCH_O += $(OBJDIR)/MinimumBayesRiskNBestListSearch.o
 LIBSPRINTSEARCH_O += $(OBJDIR)/MinimumBayesRiskSearchUtil.o
 endif
+SUBDIRS += TreeBuilder
 ifdef MODULE_SEARCH_WFST
 SUBDIRS += Wfst
 endif
@@ -58,6 +56,9 @@ libSprintSearch.$(a): $(LIBSPRINTSEARCH_O)
 
 check$(exe): $(CHECK_O)
 	$(LD) $(CHECK_O) -o check$(exe) $(LDFLAGS)
+
+TreeBuilder:
+	$(MAKE) -C $@ libSprintTreeBuilder.$(a)
 
 Wfst: 
 	$(MAKE) -C $@ libSprintSearchWfst.$(a)
